# shared/sdk-ts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IncidentActionOut": {
        "description": "Fila de ``incident_actions`` (timeline append-only del incidente).",
        "properties": {
          "action_id": {
            "format": "uuid",
            "title": "Action Id",
            "type": "string"
          },
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          }
        },
        "required": [
          "action_id",
          "incident_id",
          "tenant_id",
          "ts",
          "kind",
          "actor",
          "payload"
        ],
        "title": "IncidentActionOut",
        "type": "object"
      },
...
      "IncidentOut": {
        "description": "Fila de ``incidents`` visible por RLS (resumen para lista y detalle).",
        "properties": {
          "closed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "event_uuid": {
            "format": "uuid",
            "title": "Event Uuid",
            "type": "string"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "max_pga_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Pga G"
          },
          "max_pgv_cms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Pgv Cms"
          },
          "opened_at": {
            "format": "date-time",
            "title": "Opened At",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "site_id": {
            "format": "uuid",
            "title": "Site Id",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "event_uuid",
          "tenant_id",
          "site_id",
          "event_id",
          "opened_at",
          "closed_at",
          "severity",
          "state",
          "trigger",
          "max_pga_g",
          "max_pgv_cms",
          "summary"
        ],
        "title": "IncidentOut",
        "type": "object"
      },
      "IncidentPage": {
        "description": "Página keyset de incidentes: ``items`` + cursor opaco de continuación.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IncidentOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "IncidentPage",
        "type": "object"
      },
...
      "RuleSetList": {
        "description": "rule_sets del tenant (activos primero, luego por versión descendente).",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RuleSetOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "RuleSetList",
        "type": "object"
      },
      "RuleSetOut": {
        "description": "Fila de ``rule_sets`` visible por RLS (una versión concreta).",
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "rule_set_id": {
            "format": "uuid",
            "title": "Rule Set Id",
            "type": "string"
          },
          "scope_id": {
            "format": "uuid",
            "title": "Scope Id",
            "type": "string"
          },
          "scope_type": {
            "title": "Scope Type",
            "type": "string"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "rule_set_id",
          "tenant_id",
          "scope_type",
          "scope_id",
          "version",
          "is_active",
          "config",
          "created_by",
          "created_at"
        ],
        "title": "RuleSetOut",
        "type": "object"
      },
      "RuleSetPublishOut": {
        "description": "Respuesta 202 de publish: intención de sync al edge (real en T-1.23).",
        "properties": {
          "rule_set_id": {
            "format": "uuid",
            "title": "Rule Set Id",
            "type": "string"
          },
          "status": {
            "default": "pending_sync",
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "rule_set_id",
          "version"
        ],
        "title": "RuleSetPublishOut",
        "type": "object"
      },
      "RuleSetPutIn": {
        "description": "Alta de una NUEVA versión de rule_set para un alcance dado.",
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "scope_id": {
            "format": "uuid",
            "title": "Scope Id",
            "type": "string"
          },
          "scope_type": {
            "title": "Scope Type",
            "type": "string"
          }
        },
        "required": [
          "scope_type",
          "scope_id",
          "config"
        ],
        "title": "RuleSetPutIn",
        "type": "object"
      },
...
      "SiteDetailOut": {
        "description": "Detalle de un sitio con sus zonas.",
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "building_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Building Type"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "criticality": {
            "title": "Criticality",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "site_id": {
            "format": "uuid",
            "title": "Site Id",
            "type": "string"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "zones": {
            "items": {
              "$ref": "#/components/schemas/ZoneOut"
            },
            "title": "Zones",
            "type": "array"
          }
        },
        "required": [
          "site_id",
          "tenant_id",
          "code",
          "name",
          "timezone",
          "criticality",
          "lat",
          "lon",
          "created_at",
          "zones"
        ],
        "title": "SiteDetailOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ZoneOut": {
        "description": "Zona (planta/área) de un sitio; el polígono no se expone en el catálogo.",
        "properties": {
          "level_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level Code"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "zone_id": {
            "format": "uuid",
            "title": "Zone Id",
            "type": "string"
          }
        },
        "required": [
          "zone_id",
          "name"
        ],
        "title": "ZoneOut",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "TAKAB API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/incidents": {
      "get": {
        "description": "Lista incidentes (opened_at desc) con filtros y paginación keyset estable.",
        "operationId": "list_incidents_incidents_get",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "site_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Site Id"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Incidents"
      }
    },
    "/incidents/{incident_id}": {
      "get": {
        "description": "Detalle de un incidente. 404 si no existe o no es visible por RLS.",
        "operationId": "get_incident_incidents__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident"
      }
    },
...
    "/incidents/{incident_id}/actions": {
      "get": {
        "description": "Timeline append-only del incidente. 404 si el incidente no es visible.",
        "operationId": "list_incident_actions_incidents__incident_id__actions_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IncidentActionOut"
                  },
                  "title": "Response List Incident Actions Incidents  Incident Id  Actions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Incident Actions"
      }
    },
...
    "/rule-sets": {
      "get": {
        "description": "rule_sets del tenant (RLS): versiones activas primero.",
        "operationId": "list_rule_sets_rule_sets_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSetList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Rule Sets"
      },
      "put": {
        "description": "Crea una NUEVA versión activa del alcance (version+1) y apaga las previas.",
        "operationId": "put_rule_set_rule_sets_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleSetPutIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Rule Set"
      }
    },
    "/rule-sets/{rule_set_id}/publish": {
      "post": {
        "description": "Marca la intención de sincronizar el rule_set al edge (202 pending_sync).",
        "operationId": "publish_rule_set_rule_sets__rule_set_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "rule_set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Set Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSetPublishOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Rule Set"
      }
    },
...
    "/sites/{site_id}": {
      "get": {
        "description": "Detalle de un sitio con sus zonas. 404 si no existe o no es visible (RLS).",
        "operationId": "get_site_sites__site_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Site Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Site"
      }
    },
...
  }
}